# docs/swagger.yaml
definitions:
  api.AuctionVehicleResponse:
    properties:
      details:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/api.ImageResponse'
        type: array
      vehicle_id:
        type: integer
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.FuelingRecordResponse:
    properties:
      after_fueling_image:
        type: string
      amount:
        type: number
      before_fueling_image:
        type: string
      created_at:
        type: string
      fueling_person_id:
        type: integer
      gas_station:
        description: omitempty if the field is optional
        type: string
      id:
        type: integer
      notes:
        description: omitempty if the field is optional
        type: string
      total_cost:
        type: number
      updated_at:
        type: string
      vehicle_id:
        type: integer
    type: object
  api.ImageResponse:
    properties:
      id:
        type: integer
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This API serves as a backend for Vehicle Management System
  title: Vehicle Management System API
  version: "1.0"
paths:
  /auction:
    get:
      description: Get all auctions
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with auction details
          schema:
            items:
              $ref: '#/definitions/api.AuctionVehicleResponse'
            type: array
      summary: Get all auctions
      tags:
      - auction
    post:
      consumes:
      - multipart/form-data
      description: Admins can create auctions with vehicle details and images
      parameters:
      - description: Vehicle ID
        in: formData
        name: vehicle_id
        required: true
        type: integer
      - description: Details of the auction
        in: formData
        name: details
        type: string
      - description: Images for the auction
        in: formData
        name: images
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with auction details
          schema:
            $ref: '#/definitions/api.AuctionVehicleResponse'
      security:
      - ApiKeyAuth: []
      summary: Create an auction
      tags:
      - auction
  /auction/{id}:
    delete:
      description: Delete an auction
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with message
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete an auction
      tags:
      - auction
    get:
      description: Get an auction
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with auction details
          schema:
            $ref: '#/definitions/api.AuctionVehicleResponse'
      summary: Get an auction
      tags:
      - auction
  /fueling:
    get:
      description: Get all fueling records
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with fueling record details
          schema:
            items:
              $ref: '#/definitions/api.FuelingRecordResponse'
            type: array
        "400":
          description: Bad Request with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all fueling records
      tags:
      - fueling
    post:
      consumes:
      - multipart/form-data
      description: Admins and fueling personnel can create fueling records
      parameters:
      - description: Vehicle ID
        in: formData
        name: vehicle_id
        required: true
        type: integer
      - description: Fueling Person ID
        in: formData
        name: fueling_person_id
        required: true
        type: integer
      - description: Amount of fuel
        in: formData
        name: amount
        required: true
        type: number
      - description: Total cost of fueling
        in: formData
        name: total_cost
        required: true
        type: number
      - description: Gas Station
        in: formData
        name: gas_station
        type: string
      - description: Additional notes
        in: formData
        name: notes
        type: string
      - description: Image before fueling
        in: formData
        name: before_fueling_image
        required: true
        type: file
      - description: Image after fueling
        in: formData
        name: after_fueling_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with fueling record details
          schema:
            $ref: '#/definitions/api.FuelingRecordResponse'
        "400":
          description: Bad Request with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create fueling record
      tags:
      - fueling
  /fueling/{id}:
    delete:
      description: Delete a fueling record
      parameters:
      - description: Fueling Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with message
          schema:
            type: string
        "400":
          description: Bad Request with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a fueling record
      tags:
      - fueling
    get:
      consumes:
      - application/json
      description: Get a fueling record
      parameters:
      - description: Fueling Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with fueling record details
          schema:
            $ref: '#/definitions/api.FuelingRecordResponse'
        "400":
          description: Bad Request with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found with error message
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a fueling record
      tags:
      - fueling
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"